requirements.txt
setuptools
python-telegram-bot
qrcode[pil]